change_utils/requirements.txt
crowdin-api-client==1.24.1
pyairtable==3.3.0
python-dotenv==1.1.0
argparse==1.4.0
flatten-json==0.1.14
requests==2.32.5
google-cloud-storage
-e .